requirements.txt
streamlit>=1.28.0
pandas>=2.1.0
Pillow>=10.0.0
numpy>=1.24.0
protobuf>=4.21.0
watchdog>=3.0.0
python-dotenv==1.0.1


tqdm>=4.65.0
transformers>=4.30.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0

